kernel: add SGT allocation operation
The SVEC version allows DMA within a PAGE_SIZE (4KiB) window. This means
that we cannot perform DMA transfers bigger than PAGE_SIZE (4K).
The sg_alloc_table_from_pages automatically squash continguos transfers
to improve performances, but we cannot use it.

In order to avoid to avoid one `if`, I defined this new operation that is
configured at probe time so that we can use a no_squash version when
necessary.

Signed-off-by: Federico Vaga <federico.vaga@vaga.pv.it>